The operation to remove the 'Container' object is placed when 'dealloc' is triggered.
When using a Container, popping or dismissing it does not immediately trigger dealloc.
Additionally, if a developer wants to hold a Container object for a longer period and try to use it repeatedly, an assertion error will be triggered because the cached 'uniqueId' is removed.
This error can be avoided.